Configure for dev environment by default instead of aborting
That is, for tasks that need to be configured, assume tasks are being
run in the dev environment. That way we can do things like `inv clean`
instead of `inv dev clean`.

This makes sense to me because most tasks are run most of the time in
the dev env for dev purposes.