Add @arctask to replace Invoke's @ctask
- Use it throughout
- Remove the `configured` task
- Stop exporting ctask
- Export arctask instead

The impetus for this was to make it easier to specify that a task is
a configured task and further to make it easy to specify the default env
a task should be configured for (where having a default makes sense).

This also lets us add features like timing of tasks.