Add a way to specify default task options via config
For example, in a hypothetical module named `my.tasks`:

    @arctask(configured=True)
    def my_task(ctx, some_option=None):
        ...

In tasks.cfg:

    my.tasks.my_task.some_option = "my default value"

Now, if `my_task` is called without --some-option, --some-option will
automatically be set to "my default value".

The purpose of this is to allow per-project defaults for options that
don't have a sensible global default. This also means that tasks that
before looked up defaults in config won't have to do that any more.

To support this, had to add a custom `Config.__contains__()`.