Link version-specific static manifest when deploy()ing and link()ing
When doing a rollback using the link() task, the static manifest
(staticfiles.json) needs to be pointed at the manifest for the version
being rolled back to.

This works with the strategy of only adding new files to the remote
static directory so that clients that haven't refreshed will still be
able to access old static files.