Don't cd in remote() calls in deploy tasks
Remove cd option from remote() calls where it's not needed because
commands use abs. paths. Make rel. paths absolute in other remote()
calls so that cd'ing isn't necessary.

When running deploy() as a normal user (i.e., as myself, the default),
commands are run on the server as a service user (using `sudo -u`),
which works fine for most things, but things like cd'ing and shell
globbing won't work if my user isn't in the service user's resgrp.

Basically, this works around needing to be in a particular resgrp to do
a deployment (although, of course, this will only work for users that
have sudo).