Use first host in ALLOWED_HOSTS in restart()
Instead of using a custom `DOMAIN_NAME` setting (which should probably
have been named `HOST` or `HOSTNAME`) in the `restart()` task (for
pinging the site after a deployment), use the first entry in
`ALLOWED_HOSTS` instead.  If the first host has a leading dot, it will
be stripped.

Another thing we could do here if this doesn't work for some sites is
add a `host` option to `restart()`.