Revamp handling of default task options
Instead of setting task keyword args to None and then getting their
default values from config, task keyword arg can now be set directly.

The main advantage is that Invoke can inspect default option values so
that it can do better parsing and completion of options on the command
line.

This requires a bit of hacking around the way Invoke calls tasks--it
always passes a value for every keyword arg, even for options that
aren't specified. What it SHOULD do, in my opinion, is call tasks in
the usual way; that is, it should NOT pass keyword args to tasks for
options that aren't specified.