Revert "Fix overriding of default config with project config"
This reverts commit 41d11d0f55c2be58d10e18b0e6076cf64501448d.

The issue/question here is whether the config in a project's [DEFAULT]
section should override config in an arctasks environment section.

E.g., if prod config is being used, should the project's [DEFAULT]
config override arctasks' [prod] config?

In the original implementation, a project's [DEFAULT] config did *not*
override the arctasks env config. The "fix" in 41d11d0 inverted this.

Which of these options is correct is debatable, but the original
implementation is simpler, so it seems like a good choice.

The upshot is that if a project defines an env config section and wants
to override something in the corresponding env config section in
arctasks, the override must be in the env config section in the project
and not in the project's [DEFAULT] section.

Putting it another way, this is the config resolution order: [project
env], [arctasks env], [project DEFAULT], [arctasks DEFAULT].