configure(): save config file name for use in later calls
I.e., if configure() is called more than once, use the same config file
for the second and subsequent calls (if a config file isn't passed to
the subsequent calls).

This makes it easy to reconfigure. For example, a typical project init
task is run with dev config, but when test() is called from init, it
may be necessary or desirable to reconfigure to run with test config.