Add "default env" functionality
The `configured` option to an `arctask` can be set to `DEFAULT_ENV`
(imported from arctasks.arctask), which is a marker indicating that the
task should be run in the default env. The default default env is dev.

The default env can be set via the `ARCTASKS_DEFAULT_ENV` OS environment
variable. Note that we can't set it in the task config because...
bootstrapping.

This is intended to be used mainly when deploying to create a wrapper
around the `inv` console script that configures the default env to the
env of the deployment. The default deployer class has been updated to
do just that.

In addition, the `retrieve` task has been configured to use the default
env because A) it makes sense and B) as an example. Other tasks will
need to be reviewed to see if this change makes sense for them too.