Handle Django management commands being aborted w/ Ctrl-C
Instead of barfing out a useless stack trace, catch `KeyboardInterrupt`s
and exit with a nice message.

This would generally be applicable to management commands that stop and
ask the user for confirmation.

Note that if a management command handles Ctrl-C internally, our handler
won't have any effect (which is what we want). E.g., `runserver` catches
Ctrl-C and exits gracefully.