Don't install invoke in virtualenv task
There was a comment there saying this was necessary for bootstrapping
purposes, but that appears to be incorrect. I think this was a vestige
of an initial version of arctasks.

With this change, we also don't need to duplicate the Invoke version
in tasks.cfg. I.e., the `_invoke.version` setting was removed.