When deploying, copy & use frozen pip requirements file
I.e., we now assume that we're deploying a release that has a top level
requirements-frozen.txt file, and we tell pip to build wheels and
install using that file.

This also removes the separate `wheel` task, because I don't see any
need to run it separately, and it's not exported into the default task
set anyway.