Improve how arctasks is installed when deploying
Instead of running a separate `pip install` command on the remote host,
pull down the latest tarball to the build/local host and create a wheel
from it that gets pushed to the remote host along with other packages.

This is better because it will cause the deployment to abort early if
GitHub is down. It's also more consistent.

Eventually, arctasks should be released so that this hackery isn't
necessary.