Remove frozen requirements when resuming development
When a release is prepared, a frozen pip requirements file is generated.
When development is resumed after a release, we don't requirements to be
frozen, so we now remove the frozen requirements file when resuming
development.

We have to account for this when deploying too, so we now generate a
stub requirements file when doing non-release deployments to staging.

Incidentally, we now push the requirements file as part of pushing the
config, and we now push the config before installing on the remote host.

XXX: Maybe rethink how this works.