Revamp internals of configure task
The main change is that we now interpolate after reading each config
file and after reading options passed on the command line. Before, we
gathered all the config items and then interpolated only once, but this
doesn't always work when an item is overridden in a project config file.

There's also a bit of internal cleanup.