Fix config interpolation
The previous attempt at doing this could cause the final interpolated
settings to be incorrect in some cases because it was interpolating
eagerly before all the settings were read.

This new version reads all the settings and then does interpolation as
it used to do before the previous change. What's different is that the
interpolation logic is smarter in that it will keep interpolating until
there's nothing left to interpolate. So multiple levels of interpolation
and out-of-order interpolation now works.

Amends a1b8ac2