Make it possible to release any version
Previously, there had to be a section for the specific version being
released in the change log. Now, if such a section isn't found, we fall
back to looking for a section that has any version with "unreleased" as
the release date.

This supports, for example, the scenario where you need to do a patch
release before the next anticipated minor release.