Add SASS-compiling capabilities (add sass task)
SASS seems to be more popular than LESS these days. Bootstrap 4 uses it
(whereas previous versions used LESS), and so does Angular Material.
LibSASS also seems really fast. The downside is that configuration is
slightly more complex than for LESS because the Node version of LESS has
a nice plugin system that node-sass doesn't have.

This also updates the build_static task to detect whether the CSS
sources passed are .less or .scss files (it's possible to have both
types in a project).