Improve remote task, esp. wrt. sudo
Commands are now run on the remote server via a bash subshell. This
makes it easy to run a batch of commands as a specific user, set $PATH
just once, and ensure shell expansion happens as the correct user.

The `--many` flag was removed, since that was a hack for running
multiple commands as the same user and is no longer needed.

In addition, here docs are now used so that user-supplied commands can
use whatever quoting they need without conflict.

This is a simplified/generalized example of how remote commands are now
run:

    inv -e stage remote --cd '$HOME' --path '~/bin' 'ls * && echo ~'

    ssh -T <user>@hrimfaxi <<'EOSSH'
        sudo -u svusrNNN bash <<'EOBASH'
            cd ~ && export PATH="${PATH}:~/bin" && ls * && echo ~
    EOBASH
    EOSSH

Both `$HOME` and `~` will be expanded to svusrNNN's home directory, and
the `*` in `ls *` will be expanded with svusrNNN's permissions.