Always link() to correct static files manifest
There was a bug in link() that could cause the wrong static files
manifest to be linked. This would happen if link() was run while on
the HEAD of some branch instead of on a tag. The fix is to always use
the build dir constructed specifically for linking to a specific version
instead of using the build dir from the task config.