Rewrite sass task to use temp files
The previous version used pipes, but those don't seem to work well for
large files, sometimes causing output to be truncated. An example of
this is when the Angular Material SCSS file is imported.

The new version saves the output of each processing step to a temp file
and then copies the last temp file to the destination. Temp files are
cleaned up automatically.

This also cleans up a couple other things internally and prints out
what it's doing when Invoke's --echo flag is set.