Revamp how database creation (and related) commands are run
When creating databases, always run commands by passing SQL statements
to the server. In particular, this means not using Postgres's command
line tools such as createdb and dropdb. This makes things consistent and
allows for some internal cleanup to remove redundancy.

In addition, always use the database's default superuser (postgres or
root) to run create and drop commands instead of relying on the current
OS user existing as a DB superuser.