Switch from Invoke to TaskRunner
Invoke has some long-standing issues and seems to be only sporadically
developed. TaskRunner has all the features we need.

Most of the changes are just changing imports and search-and-replace of
`ctx` with `config`.

A few other changes were required to how Django stuff is set up, but
that was relatively minor and is a better organization/architecture
regardless.